Modularize React components and temporarily change index.html to not load main.js to speed development testing.
