+++ index.html
@@ -193,7 +193,7 @@
     <script src="/js/ejs_production.js" type="text/javascript"></script>
     <script src="/js/chrome-u2f-api.js" type="text/javascript"></script>
     <script src="/js/ledger.min.js" type="text/javascript"></script>
-    <script src="/js/main.js" type="text/javascript"></script>
+    <!-- TEMP: <script src="/js/main.js" type="text/javascript"></script>-->
     <!-- Must be loaded after main.js populates the redux EtherDelta.store -->
     <script src="/migrate-tool/migrate_libraries.js" type="text/javascript"></script>
     <script src="/migrate-tool/migrate_bundle.js" type="text/javascript"></script>
