+++ migrate-tool/migrate.js
@@ -46,570 +46,11 @@ try {
 }
 //TEMP============================================================================
 
-// Parent React component that controls window state
-class MigrationTool extends React.Component {
-    constructor(props) {
-        super(props);
-        this.state = {
-            // Initialize window to -3 (LoadingWindow)
-            user_address: '',
+import Application from './components/Application';
 
-            window: -3,
-
-            balances_fetched: false,
-            orders_fetched: false,
-
-            balances_options: [], // Addresses of tokens with balances for multi-select
-            balances_selected: [], // Addresses of tokens selected by user in multi-select
-
-            tokens_added: [], // Token addresses manually added by users
-
-            orders_options: [],
-            orders_selected: []
-        };
-    };
-
-    /* --- Navigate to next window -- */
-    nextWindow(change_from, change_to) {
-        // If window change_from is balances or orders, remember state of multiselect
-        this.save_ms_state(change_from, change_to);
-        this.load_ms_state(change_to);
-
-        // Change to next window
-        let current_window = this.state.window;
-        if (current_window < 6) {
-            current_window += 1;
-            this.setState({
-                window: current_window
-            });
-        }
-    };
-
-    /* --- Navigate to previous window -- */
-    previousWindow(change_from, change_to) {
-        // If window change_from is balances or orders, remember state of multiselect
-        this.save_ms_state(change_from);
-        this.load_ms_state(change_to);
-
-        // Change to previous window
-        let current_window = this.state.window;
-        if (current_window > 1) {
-            current_window -= 1;
-            this.setState({
-                window: current_window
-            });
-        }
-    };
-
-    /* --- Hides window --- */
-    closeWindow() {
-        document.getElementById('migrationTool').style.display = 'none';
-    }
-
-    /* --- Add user-entered token to selected list --- */
-    add_token() {
-        console.log("Fetching information about user entered-token not supported by FD.");
-        // TODO: Fix sloppy practice of using jQuery to access react-rendered element
-        // Get token address from element
-        let address = $('#new-token-addr-input').val().toString();
-
-        // Check that token isn't already being displayed
-        let already_displayed = false;
-        let existing_options = this.state.balances_options.concat(this.state.balances_selected);
-        existing_options.forEach(function (existing_addr) {
-            if (existing_addr === address) {
-                alert("Token balance is already listed.");
-                already_displayed = true;
-            }
-        });
-
-        // Add token to selected and tokens_added in state for updating React component
-        if (!already_displayed) {
-            let new_tokens_added = this.state.tokens_added.concat(address);
-            let new_balances_selected = this.state.balances_selected.concat(address);
-            this.setState({
-                balances_selected: new_balances_selected,
-                tokens_added: new_tokens_added
-            });
-        }
-    }
-
-    update_orders() {
-        // Check that balances are only fetched once
-        if (!this.state.orders_fetched) {
-            let temp_addresses = get_orders_options();
-            this.setState({
-                orders_options: temp_addresses,
-                orders_fetched: true
-            });
-        }
-    }
-
-    /* --- Populates balances_options with fetched balances --- */
-    update_balances() {
-        // Check that balances are only fetched once
-        if (!this.state.balances_fetched) {
-            let temp_addresses = get_balances_options();
-            this.setState({
-                balances_options: temp_addresses,
-                balances_fetched: true
-            });
-        }
-    }
-
-    /* --- Update or initialize balances multiselect --- */
-    update_balances_ms() {
-        $('#ED-migration-balances').multiselect({
-                search: {
-                    left: '<input type="text" name="ED-bal-search" class="form-control" placeholder="Search Ether Balances" />',
-                    right: '<input type="text" name="FD-bal-search" class="form-control" placeholder="Search Fork Balances" />'
-                },
-                fireSearch: function (value) {
-                    return value.length > 1;
-                }
-        });
-    }
-
-    /* --- Update or initialize orders multiselect --- */
-    update_orders_ms() {
-        $('#ED-migration-orders').multiselect({
-            search: {
-                left: '<input type="text" name="ED-bal-search" class="form-control" placeholder="Search Ether Orders" />',
-                right: '<input type="text" name="FD-bal-search" class="form-control" placeholder="Search Fork Orders" />'
-            },
-            fireSearch: function (value) {
-                return value.length > 1;
-            }
-        });
-    }
-
-    /* --- Get the option values in a given multiselect --- */
-    get_ms_selected_values(ms_id) {
-        let selected = [];
-        $(ms_id + ' option').each(function () {
-            selected.push($(this).val());
-        });
-        return selected;
-    }
-
-    save_ms_state(change_from) {
-        // If window change_from is balances, save state of multiselect
-        if (change_from === "balances") {
-            let remaining_balances = this.get_ms_selected_values('#ED-migration-balances');
-            let selected_balances = this.get_ms_selected_values('#ED-migration-balances_to');
-            this.setState({
-                balances_options: remaining_balances,
-                balances_selected: selected_balances
-            })
-        }
-
-        // If window change_from is orders, save state of multiselect
-        if (change_from === "orders") {
-            let remaining_orders = this.get_ms_selected_values('#ED-migration-orders');
-            let selected_orders = this.get_ms_selected_values('#ED-migration-orders_to');
-            this.setState({
-                orders_options: remaining_orders,
-                orders_selected: selected_orders
-            });
-        }
-    }
-
-    load_ms_state(change_to) {
-        // If window change_from is balances, load state of multiselect
-        if (change_to === "balances") {;
-            this.update_balances_ms();
-        }
-
-        // If window change_from is orders, load state of multiselect
-        if (change_to === "orders") {
-            this.update_orders_ms();
-        }
-    }
-
-    get_user_address() {
-        let user_address_temp = get_user_address();
-        this.setState({
-            user_address: user_address_temp
-        })
-    }
-
-    render() {
-        // Determine current window
-        let current_window;
-        switch (this.state.window) {
-            case -3:
-                current_window = <LoadingWindow/>;
-                break;
-            case -2:
-                current_window = <NoAccountWindow closeWindow={() => this.closeWindow()}/>;
-                break;
-            case -1:
-                current_window = <NoBalancesWindow closeWindow={() => this.closeWindow()}/>;
-                break;
-            case 1:
-                current_window = <IntroductionWindow nextWindow={() => this.nextWindow('', 'balances')}
-                                                     get_user_address={() => this.get_user_address()}
-                                                     user_address={this.state.user_address}/>;
-                break;
-            case 2:
-                current_window = <BalancesWindow nextWindow={() => this.nextWindow("balances", '')}
-                                                 previousWindow={() => this.previousWindow("balances", '')}
-                                                 balances_options={this.state.balances_options}
-                                                 balances_selected={this.state.balances_selected}
-                                                 user_address={this.state.user_address}
-                                                 update_balances={() => this.update_balances()}
-                                                 update_balances_ms={() => this.update_balances_ms()}/>;
-                break;
-            case 3:
-                current_window = <AddBalanceWindow nextWindow={() => this.nextWindow('', 'orders')}
-                                                   previousWindow={() => this.previousWindow('', 'balances')}
-                                                   add_token={() => this.add_token()}
-                                                   tokens_added={this.state.tokens_added}
-                                                   user_address={this.state.user_address}/>;
-                break;
-            case 4:
-                current_window = <OrdersWindow nextWindow={() => this.nextWindow("orders", '')}
-                                               previousWindow={() => this.previousWindow("orders", '')}
-                                               orders_options={this.state.balances_options}
-                                               orders_selected={this.state.orders_selected}
-                                               user_address={this.state.user_address}
-                                               update_orders={() => this.update_orders()}
-                                               update_orders_ms={() => this.update_orders_ms()}/>;
-                break;
-            case 5:
-                current_window = <ConfirmationWindow nextWindow={() => this.nextWindow()}
-                                                     previousWindow={() => this.previousWindow('', 'orders')}/>;
-                break;
-            case 6:
-                current_window = <SuccessWindow closeWindow={() => this.closeWindow()}/>;
-                break;
-        }
-
-        return (
-            // Modal body, with interchangeable nested window
-            <div className="migration-overlay" id="migration-modal">
-                <div className="modal-dialog" tabIndex="-1">
-                    {current_window}
-                </div>
-            </div>
-        )
-
-    };
-}
-
-class LoadingWindow extends React.Component {
-    render() {
-        return (<div className="migration-loader">
-            <i className="fa fa-spinner fa-pulse"></i>
-        </div>)
-    }
-}
-
-class NoAccountWindow extends React.Component {
-    render() {
-        return (
-            <div className="modal-content">
-                <div className="modal-header text-center">
-                    <h2>Migration Tool</h2>
-                    <p>From The EtherDelta's Smart Contract to the Improved ForkDelta Smart Contract </p>
-                </div>
-                <div className="modal-body text-center">
-                    <p><strong> Please select an account to use the migration tool. </strong></p>
-                    <br />
-                    <button className="btn btn-default" onClick={this.props.closeWindow}>Close Tool</button>
-                </div>
-            </div>
-        )
-    }
-}
-
-class NoBalancesWindow extends React.Component {
-    render() {
-        return (
-            <div className="modal-content">
-                <div className="modal-header text-center">
-                    <h2>Migration Tool</h2>
-                    <p>From The EtherDelta's Smart Contract to the Improved ForkDelta Smart Contract </p>
-                </div>
-                <div className="modal-body text-center">
-                    <p><strong> Your account has no balances on the old EtherDelta Smart Contract!</strong></p>
-                    <p> Your selected account has no need of this migration tool. </p>
-                    <p> If you'd like to check another account, return to ForkDelta and select a different account in the top-right. </p>
-                    <br />
-                    <button className="btn btn-default" onClick={this.props.closeWindow}>Return to ForkDelta</button>
-                </div>
-            </div>
-        )
-    }
-}
-
-class IntroductionWindow extends React.Component {
-    constructor(props) {
-        super(props);
-        this.props.get_user_address();
-    }
-    render() {
-        let user_addr = this.props.user_address;
-        return (
-            <div className="modal-content">
-                <div className="modal-header text-center">
-                    <h2>Migration Tool</h2>
-                    <p>From The EtherDelta's Smart Contract to the Improved ForkDelta Smart Contract </p>
-                </div>
-                <div className="modal-body text-center">
-                    <p>This tool is currently checking the following account address: {user_addr}</p>
-                    <p><strong> * Describe the tool and why the transfer is needed here * </strong></p>
-                    <br />
-                    <button className="btn btn-default" onClick={this.props.nextWindow}>Proceed</button>
-                </div>
-            </div>
-        )
-    }
-}
-
-class BalancesWindow extends React.Component {
-    constructor(props) {
-        super(props);
-        this.props.update_balances();
-        this.props.update_balances_ms();
-    }
-
-    componentDidUpdate() {
-        this.props.update_balances_ms();
-    }
-
-    render() {
-        //Map balances to options
-        let balances_options_selects = this.props.balances_options.map((token_addr) => {
-            let name = get_token_symbol(token_addr).toString();
-            let balance = get_token_balance(token_addr, this.props.user_address).toString();
-            return (
-                <option key={token_addr} value={token_addr}>{name + ":" + balance}</option>
-            )
-        });
-
-        // Map selected balances to selected
-        let balances_selected_selects = this.props.balances_selected.map((token_addr) => {
-            let name = get_token_symbol(token_addr).toString();
-            let balance = get_token_balance(token_addr, this.props.user_address).toString();
-            return (
-                <option key={token_addr} value={token_addr}>{name + ":" + balance}</option>
-            )
-        });
-
-        return (
-            <div className="modal-content">
-                <div className="modal-header text-center">
-                    <h2>Balances</h2>
-                    <p>Currently on the EtherDelta Contract to be Transferred to the ForkDelta Contract </p>
-                </div>
-                <div className="modal-body text-center">
-                    <div className="container migration-container">
-                        <div className="row">
-                            <div className="col-sm-5">
-                                <label htmlFor="ED-migration-balances">Currently on ED Contract</label>
-                                <select name="ED-bal-from[]" id="ED-migration-balances" className="form-control migrate-ms"
-                                        size="8" multiple="multiple">{balances_options_selects}</select>
-                            </div>
-                            <div className="col-sm-2 ms-buttons">
-                                <button type="button" id="ED-migration-balances_rightAll" className="btn btn-block"><i
-                                    className="fa fa-angle-double-right"></i></button>
-                                <button type="button" id="ED-migration-balances_rightSelected" className="btn btn-block"><i
-                                    className="fa fa-angle-right"></i></button>
-                                <button type="button" id="ED-migration-balances_leftSelected" className="btn btn-block"><i
-                                    className="fa fa-angle-left"></i></button>
-                                <button type="button" id="ED-migration-balances_leftAll" className="btn btn-block"><i
-                                    className="fa fa-angle-double-left"></i></button>
-                            </div>
-                            <div className="col-sm-5">
-                                <label htmlFor="ED-migration-balances_to">Pending Transfer To FD Contract</label>
-                                <select name="ED-bal-to[]" id="ED-migration-balances_to" className="form-control migrate-ms"
-                                        size="8" multiple="multiple">{balances_selected_selects}</select>
-                            </div>
-                        </div>
-                    </div>
-                    <br />
-                    <button className="btn btn-default" onClick={this.props.previousWindow}>Previous (Introduction)</button>
-                    <button className="btn btn-default" onClick={this.props.nextWindow}>Next (Add Token)</button>
-                </div>
-            </div>
-        )
-    }
-}
-
-class AddBalanceWindow extends React.Component {
-    constructor(props) {
-        super(props);
-
-        this.handleSubmit = this.handleSubmit.bind(this);
-    }
-
-    handleSubmit(event) {
-        event.preventDefault();
-        this.props.add_token();
-    }
-
-    mapAddedTokens() {
-        // If no tokens, display "No Tokens Added"
-        if (this.props.tokens_added.length < 1) {
-            return (<li key={"None"} className="list-group-item">No Tokens Added</li>)
-        }
-
-        // Map added tokens to list items
-        return this.props.tokens_added.map((token_addr) => {
-            // If no added tokens, return list item showing none added
-            let name = get_token_symbol(token_addr).toString();
-            if (name === "Name Error") {
-                name = "EnteredToken"
-            }
-            let balance = get_token_balance(token_addr, this.props.user_address).toString();
-            return (
-                <li key={token_addr} className="list-group-item">
-                    <input value={token_addr} type="checkbox" defaultChecked/>{name + ":" + balance}
-                </li>
-            )
-        });
-    }
-
-    render() {
-        let tokens_added_list= this.mapAddedTokens();
-
-        return (
-            <div className="modal-content">
-                <div className="modal-header text-center">
-                    <h2>Missing A Balance?</h2>
-                    <p>Add Tokens Not Listed With ForkDelta's Tokens</p>
-                </div>
-                <div className="modal-body text-center">
-                    <form id="token-addr-form" onSubmit={this.handleSubmit}>
-                        <div className="form-group">
-                            <label htmlFor="new-token-addr-input">Token Address</label>
-                            <input required size="42" pattern="^0x.{40}$" className="form-control"
-                                   id="new-token-addr-input"
-                                   placeholder="0x0000000000000000000000000000000000000000" />
-                        </div>
-                        <button type="submit" className="btn btn-default">Add token</button>
-                    </form>
-                    <h3> Added tokens </h3>
-                    <ul className="list-group">
-                        {/*TODO: Style tokens list*/}
-                        {/*TODO: Loading icon on add new token*/}
-                        {tokens_added_list}
-                    </ul>
-                    <br />
-                    <button className="btn btn-default" onClick={this.props.previousWindow}>Previous (Balances)</button>
-                    <button className="btn btn-default" onClick={this.props.nextWindow}>Next (Orders)</button>
-                </div>
-            </div>
-        )
-    }
-}
-
-class OrdersWindow extends React.Component {
-    constructor(props) {
-        super(props);
-        this.props.update_orders();
-        this.props.update_orders_ms();
-    }
-
-    componentDidUpdate() {
-        this.props.update_orders_ms();
-    }
-
-    render() {
-        return (
-            <div className="modal-content">
-                <div className="modal-header text-center">
-                    <h2>Orders</h2>
-                    <p>Currently on the EtherDelta Contract to be Transferred to the ForkDelta Contract </p>
-                </div>
-                <div className="modal-body text-center">
-                    <div className="container migration-container">
-                        <div className="row">
-                            <Multiselect />
-                        </div>
-                    </div>
-                    <br />
-                    <button className="btn btn-default" onClick={this.props.previousWindow}>Previous (Add Token)</button>
-                    <button className="btn btn-default" onClick={this.props.nextWindow}>Next (Confirmation)</button>
-                </div>
-            </div>
-        )
-    }
-}
-
-class ConfirmationWindow extends React.Component {
-    constructor(props) {
-        super(props);
-    }
-
-    mapAllSelected() {
-        // If no tokens, display "No Tokens Added"
-        if (this.props.tokens_added.length < 1) {
-            return (<li key={"None"} className="list-group-item">Nothing Selected for Migration...</li>)
-        }
-
-        // Map added tokens to list items
-        return this.props.tokens_added.map((token_addr) => {
-            // If no added tokens, return list item showing none added
-            let name = get_token_symbol(token_addr).toString();
-            if (name === "Name Error") {
-                name = "EnteredToken"
-            }
-            let balance = get_token_balance(token_addr, this.props.user_address).toString();
-            return (
-                <li key={token_addr} className="list-group-item">
-                    <input value={token_addr} type="checkbox" defaultChecked/>{name + ":" + balance}
-                </li>
-            )
-        });
-    }
-
-    render() {
-        let all_selected = this.all_selected
-        return (
-            <div className="modal-content">
-                <div className="modal-header text-center">
-                    <h2>Confirmation</h2>
-                </div>
-                <div className="modal-body text-center">
-                    <p>If all the information is correct, begin </p>
-                    <h3> Estimated Gas Cost </h3>
-                    <h3> Added tokens </h3>
-                    <ul className="list-group">
-                        {/*TODO: Style list*/}
-                        {all_selected}
-                    </ul>
-                    <br />
-                    <button className="btn btn-default" onClick={this.props.previousWindow}>Previous (Orders)</button>
-                    <button className="btn btn-default" onClick={this.props.nextWindow}>Begin Migration</button>
-                </div>
-            </div>
-        )
-    }
-}
-
-class SuccessWindow extends React.Component {
-    render() {
-        return (
-
-            <div className="modal-content">
-                <div className="modal-header text-center">
-                    <h2>Progress</h2>
-                    <p>You Can Close This Window and Check Your Progress At Any Time</p>
-                </div>
-                <div className="modal-body text-center">
-                    <p> Progress </p>
-                    <br />
-                    <button className="btn btn-default" onClick={this.props.closeWindow}>Close Migration Tool</button>
-                </div>
-            </div>
-        )
-    }
-}
 
 // React component initialization for front-end of migration tool
-var react_component = ReactDOM.render(<MigrationTool/>, document.getElementById('migrationTool'));
+var react_component = ReactDOM.render(<Application/>, document.getElementById('migrationTool'));
 
 
 // populate token info arrays from JSON
@@ -664,10 +105,11 @@ var user_address = '0xa83adca55ce5d0cc43ba16f4247ca65229a1bfb1';
 // var user_address;
 
 // call async function to load user address
-var user_address_promise = fetch_user_addr();
+// TEMP: Just loading user address
+// var user_address_promise = fetch_user_addr();
 
 // once user address loads, check for ED balances and run migration tool if any balances exist for account
-user_address_promise.then(function (promised_user_addr) {
+// user_address_promise.then(function (promised_user_addr) {
 
     // TEMP: change promised_user_addr to user_address to get browser user address
     // user_address = promised_user_addr;
@@ -689,10 +131,10 @@ user_address_promise.then(function (promised_user_addr) {
         change_window(1);
 
     } else {
-        // Change window to NoBalancesWindow
+        // Change window to NoBalancesWindow.js
         change_window(-1);
     }
-});
+// });
 
 // Front-end event listeners
 // nav-bar link to open migration tool
